+++ packages/dara-core/dara/core/interactivity/data_variable.py
@@ -182,6 +182,7 @@ class DataVariable(AnyDataVariable):
         store: CacheStore,
         filters: Optional[Union[FilterQuery, dict]] = None,
         pagination: Optional[Pagination] = None,
+        format_for_display: bool = False,
     ) -> Optional[DataFrame]:
         """
         Get the value of this DataVariable.
@@ -212,6 +213,12 @@ class DataVariable(AnyDataVariable):
 
         if entry.data is not None:
             filtered_data, count = apply_filters(entry.data, coerce_to_filter_query(filters), pagination)
+            if format_for_display and filtered_data is not None:
+                filtered_data = filtered_data.copy()
+                for col in filtered_data.columns:
+                    if filtered_data[col].dtype == 'object':
+                        # We need to convert all values to string to avoid issues with displaying data in the Table component, for example when displaying datetime and number objects in the same column
+                        filtered_data.loc[:, col] = filtered_data[col].apply(str)
             data = filtered_data
             # Store count for given filters and schema
             await asyncio.gather(